Change some types and fix some harmless signedness warnings.

+++ nk/io.h
@@ -31,9 +31,9 @@
 
 #include <sys/socket.h>
 
-int safe_read(int fd, char *buf, int len);
-int safe_write(int fd, const char *buf, int len);
-int safe_sendto(int fd, const char *buf, int len, int flags,
-                const struct sockaddr *dest_addr, socklen_t addrlen);
+ssize_t safe_read(int fd, char *buf, size_t len);
+ssize_t safe_write(int fd, const char *buf, size_t len);
+ssize_t safe_sendto(int fd, const char *buf, size_t len, int flags,
+                    const struct sockaddr *dest_addr, socklen_t addrlen);
 
 #endif /* NCM_IO_H_ */
